split out util
